// types.ts
export interface Milestone {
  year: string;
  role: string;
  organization: string;
  description: string;
}

export interface Achievement {
  title: string;
  category: string;
  summary: string;
}

export interface Initiative {
  id: string;
  title: string;
  description: string;
  image: string;
}